template info.json and syntax

+++ src/template/v2/template.mk
@@ -10,11 +10,15 @@ BUILD    			= build
 # RM
 RM 					= rm
 
+__ASCAN_BEGIN__
+ascan-if --color and no --no-color:
 # COLORS
 CLR_RED				= $(shell tput setaf 1)
 CLR_GREEN			= $(shell tput setaf 2)
 CLR_YELLOW			= $(shell tput setaf 3)
 CLR_RESET			= $(shell tput sgr0)
+ascan-fi
+__ASCAN_END__
 
 # TARGETS
 __ASCAN_BEGIN__
@@ -25,7 +29,8 @@ __ASCAN_END__
 
 # ==============================================================================
 # RULES
-
+__ASCAN_BEGIN__
+ascan-if --mode-control and no --no-mode-control:
 default: debug
 
 debug: CXXFLAGS += -g -DDEBUG=1
@@ -33,6 +38,10 @@ debug: $(BUILD)/debug.mode __ASCAN_BEGIN__ python3 targets.py list $(ascan.targe
 
 release: CXXFLAGS += -O3# -DNDEBUG=1
 release: $(BUILD)/release.mode __ASCAN_BEGIN__ python3 targets.py list $(ascan.targets) __ASCAN_END__
+ascan-else:
+default: __ASCAN_BEGIN__ python3 targets.py list $(ascan.targets) __ASCAN_END__
+ascan-fi
+__ASCAN_END__
 
 # ==============================================================================
 # EXECUTABLE DETAILS
@@ -53,7 +62,7 @@ __ASCAN_BEGIN__
 # 	@printf "$(CLR_GREEN)Linking $@...\n$(CLR_RESET)"
 # 	@printf "$(CXX) $(CXXFLAGS) -o $@ ... $(LFLAGS)\n"
 # 	@$(CXX) $(CXXFLAGS) -o $@ $^ $(LFLAGS)
-python3 executable.py $(ascan.targets) ; $(ascan.sources)
+python3 executable.py $(ascan.args) ; $(ascan.targets) ; $(ascan.sources)
 __ASCAN_END__
 
 # ==============================================================================
@@ -62,10 +71,16 @@ __ASCAN_END__
 clean:
 	$(RM) -rf "$(BUILD)" __ASCAN_BEGIN__ python3 targets.py list $(ascan.targets) __ASCAN_END__
 
+__ASCAN_BEGIN__
+ascan-if --mode-control and no --no-mode-control:
 .PHONY: default 	\
 		debug 		\
 		release 	\
 		clean
+ascan-else:
+.PHONY: default clean
+ascan-fi
+__ASCAN_END__
 
 .SECONDARY: $(BUILD)/%.d $(BUILD)/%.o
 
@@ -73,6 +88,8 @@ clean:
 #         YOU PROBABLY DON'T WANT TO MODIFY ANYTHING BELOW THIS LINE		   #
 ################################################################################
 
+__ASCAN_BEGIN__
+ascan-if --mode-control and no --no-mode-control:
 # ==============================================================================
 # MODE CONTROL
 
@@ -81,9 +98,13 @@ MODE_FILE = $(BUILD)/mode.lock
 $(BUILD)/%.mode:
 	@if [ ! -f "$@" ]; then 											\
 		if [ -f $(MODE_FILE) ]; then 									\
+		ascan-if --color and no --no-color:
 			printf "$(CLR_GREEN)"; 										\
+		ascan-fi
 			printf "Switching to $* mode, rebuilding...\n"; 			\
+		ascan-if --color and no --no-color:
 			printf "$(CLR_RESET)"; 										\
+		ascan-fi
 			$(RM) -f "$(BUILD)"/*.mode;									\
 			touch $(MODE_FILE) "$@"; 									\
 			$(MAKE) --no-print-directory -s depend; 					\
@@ -93,6 +114,8 @@ $(BUILD)/%.mode:
 			touch $(MODE_FILE) "$@"; 									\
 		fi; 															\
 	fi
+ascan-fi
+__ASCAN_END__
 
 # ==============================================================================
 # COMPILE OBJECT FILES
